Move ONNX/TensorRT export to concrete DINOv3 LTDETR model
The export_onnx and export_tensorrt functions, plus the
get_export_output_names hook, are task-specific (e.g. the ONNX
verification collapses the query dimension, which only makes sense for
detection outputs). Move them from _DINOv3LTDETRBase to the concrete
DINOv3LTDETRObjectDetection model so the base stays task-agnostic.
