+++ src/lightly_train/_task_models/dinov3_ltdetr/task_model.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Sequence
-from copy import deepcopy
 from typing import Any, Literal, NoReturn, cast
 
 import torch
@@ -19,15 +18,8 @@ from torch import Tensor, nn
 from torchvision.transforms.v2 import functional as transforms_functional
 from typing_extensions import Self
 
-from lightly_train import _logging, _torch_testing
-from lightly_train._commands import _warnings
 from lightly_train._configs.config import PydanticConfig
 from lightly_train._data import file_helpers
-from lightly_train._export import tensorrt_helpers
-from lightly_train._export.onnx_helpers import (
-    fix_topological_order,
-    remove_redundant_casts,
-)
 from lightly_train._models import package_helpers
 from lightly_train._models.dinov3.dinov3_convnext import DINOv3VConvNeXtModelWrapper
 from lightly_train._models.dinov3.dinov3_package import DINOV3_PACKAGE
@@ -844,9 +836,6 @@ class _DINOv3LTDETRBase(TaskModel):
     def build_postprocessor(self, config: _DINOv3LTDETRConfig) -> nn.Module:
         raise NotImplementedError()
 
-    def get_export_output_names(self) -> list[str]:
-        raise NotImplementedError()
-
     def freeze_backbone(self) -> None:
         self.backbone.eval()
         self.backbone.requires_grad_(False)
@@ -971,320 +960,4 @@ class _DINOv3LTDETRBase(TaskModel):
         raw = self.forward_backend(batch)
         return cast(
             dict[str, Tensor], self.postprocess(raw, [metadata], threshold=threshold)[0]
-        )
-
-    @torch.no_grad()
-    def export_onnx(
-        self,
-        out: PathLike,
-        *,
-        precision: Literal["fp32", "fp16"] = "fp32",
-        batch_size: int = 1,
-        dynamic_batch_size: bool = True,
-        opset_version: int | None = None,
-        simplify: bool = True,
-        verify: bool = True,
-        format_args: dict[str, Any] | None = None,
-        num_channels: int | None = None,
-    ) -> None:
-        """Exports the model to ONNX for inference.
-
-        The export uses a dummy input of shape (batch_size, C, H, W) where C is
-        inferred from the first model parameter and (H, W) come from
-        `self.image_size`. If `dynamic_batch_size` is True, the ONNX graph will
-        have a dynamic batch dimension for the input. The graph output names are provided by the concrete task model.
-
-        Optionally simplifies the exported model in-place using onnxslim and
-        verifies numerical closeness against a float32 CPU reference via
-        ONNX Runtime.
-
-        Args:
-            out:
-                Path where the ONNX model will be written.
-            precision:
-                Precision for the ONNX model. Either "fp32", or "fp16".
-            batch_size:
-                Batch size for the ONNX input.
-            dynamic_batch_size:
-                If True, the ONNX graph will have a dynamic batch dimension for the
-                input. If False, the batch dimension is fixed to `batch_size`.
-            opset_version:
-                ONNX opset version to target. If None, PyTorch's default opset is used.
-            simplify:
-                If True, run onnxslim to simplify and overwrite the exported model.
-            verify:
-                If True, validate the ONNX file and compare outputs to a float32 CPU
-                reference forward pass.
-            format_args:
-                Optional extra keyword arguments forwarded to `torch.onnx.export`.
-            num_channels:
-                Number of input channels. If None, will be inferred.
-
-        Returns:
-            None. Writes the ONNX model to `out`.
-
-        """
-        # Set up logging.
-        _warnings.filter_export_warnings()
-        _logging.set_up_console_logging()
-
-        # Set the model in eval and deploy mode.
-        self.eval()
-
-        if precision not in ("fp32", "fp16"):
-            raise ValueError(
-                f"Invalid precision '{precision}'. Must be one of 'fp32', 'fp16'."
-            )
-
-        # Always trace in fp32 to avoid dtype mismatches in the decoder's
-        # autocast(enabled=False) blocks. fp16 conversion is applied
-        # post-export via onnxruntime.transformers.
-        self.to(torch.float32)
-        self.deploy()
-        model_device = next(self.parameters()).device
-
-        # Try to infer num_channels if not provided.
-        if num_channels is None:
-            if self.image_normalize is not None:
-                num_channels = len(self.image_normalize["mean"])
-                logger.info(
-                    f"Inferred num_channels={num_channels} from image_normalize."
-                )
-            else:
-                # Try to find the number of channels from the first convolutional layer.
-                for module in self.modules():
-                    if isinstance(module, torch.nn.Conv2d):
-                        num_channels = module.in_channels
-                        logger.info(
-                            f"Inferred num_channels={num_channels} from first Conv. layer."
-                        )
-                        break
-                if num_channels is None:
-                    logger.error(
-                        "Could not infer num_channels. Please provide it explicitly."
-                    )
-                    raise ValueError(
-                        "num_channels must be provided for ONNX export if it cannot be inferred."
-                    )
-
-        if dynamic_batch_size:
-            batch_size = 2
-        dynamic_axes = {"images": {0: "N"}} if dynamic_batch_size else None
-
-        # Create dummy input using same device and dtype as the model.
-        dummy_input = torch.randn(
-            batch_size,
-            num_channels,
-            self.image_size[
-                0
-            ],  # TODO(Thomas, 12/25): Allow passing different image size.
-            self.image_size[1],
-            requires_grad=False,
-            device=model_device,
-            dtype=torch.float32,
-        )
-
-        # TODO(Thomas, 12/25): Add warm-up forward if needed.
-
-        # Set the input/output names.
-        input_names = ["images"]
-        output_names = self.get_export_output_names()
-
-        # TODO(Nauryzbay, 05/2026): When refactoring forward() to use forward_backend(),
-        # expose orig_target_size as a second ONNX input to rescale boxes to original
-        # image coordinates inside the graph.
-        torch.onnx.export(
-            self,
-            (dummy_input,),
-            str(out),
-            input_names=input_names,
-            output_names=output_names,
-            opset_version=opset_version,
-            dynamo=False,
-            dynamic_axes=dynamic_axes,
-            **(format_args or {}),
-        )
-
-        if precision == "fp16":
-            # convert_float_to_float16 creates nodes with duplicate names. In order to avoid downstream issues
-            # we require simplify to be True, as this correctly renames nodes.
-            if not simplify:
-                raise ValueError("fp16 precision requires simplify=True.")
-
-            import onnx
-            from onnxruntime.transformers import float16 as ort_float16
-
-            model_onnx = onnx.load(str(out))
-            # If the input to Softmax are too large the output of Softmax will be NaN values. Therefore we run
-            #  the Softmax computation in fp32. The nodes before Softmax are always MatMul.
-            # TODO (simon, 05/26) Ideally we would only block operators were a Matmul directly feeds into a Softmax.
-            op_block_list = list(ort_float16.DEFAULT_OP_BLOCK_LIST) + [
-                "Softmax",
-                "MatMul",
-            ]
-            model_fp16 = ort_float16.convert_float_to_float16(
-                model_onnx, op_block_list=op_block_list
-            )
-            # Using the op blocklist on a graph that looks like Softmax -> MatMul creates a graph that looks like
-            #  Cast32 -> MatMul -> Cast16 -> Cast32 -> Softmax -> Cast16. Therefore, we need to remove the middle
-            #  Cast16 -> Cast32.
-            remove_redundant_casts(model_fp16)
-            fix_topological_order(model_fp16)
-            onnx.save(model_fp16, str(out))
-
-        if simplify:
-            import onnxslim  # type: ignore [import-not-found,import-untyped]
-
-            onnxslim.slim(
-                str(out),
-                output_model=out,
-            )
-
-        if verify:
-            logger.info("Verifying ONNX model")
-            import onnx
-            import onnxruntime as ort
-
-            onnx.checker.check_model(out, full_check=True)
-
-            providers = ort.get_available_providers()
-            if precision == "fp16" and "CUDAExecutionProvider" not in providers:
-                logger.warning(
-                    "Skipping ONNX runtime verification for fp16 model because "
-                    "CUDAExecutionProvider is not available in onnxruntime. "
-                    "Install onnxruntime-gpu to enable full verification."
-                )
-            else:
-                # Always run the reference input in float32 and on cpu for consistency.
-                reference_model = deepcopy(self).cpu().to(torch.float32).eval()
-                reference_model.deploy()
-                reference_outputs = reference_model(
-                    dummy_input.cpu().to(torch.float32),
-                )
-
-                # Get outputs from the ONNX model. Load from bytes to avoid
-                # ORT errors about missing external data when weights are inline.
-                with open(out, "rb") as f:
-                    session = ort.InferenceSession(f.read())
-                onnx_input = dummy_input.cpu()
-                if precision == "fp16":
-                    onnx_input = onnx_input.half()
-                input_feed = {
-                    "images": onnx_input.numpy(),
-                }
-                outputs_onnx = session.run(output_names=None, input_feed=input_feed)
-                outputs_onnx = tuple(torch.from_numpy(y) for y in outputs_onnx)
-
-                # Verify that the outputs from both models are close.
-                if len(outputs_onnx) != len(reference_outputs):
-                    raise AssertionError(
-                        f"Number of onnx outputs should be {len(reference_outputs)} but is {len(outputs_onnx)}"
-                    )
-                for output_onnx, output_model, output_name in zip(
-                    outputs_onnx, reference_outputs, output_names
-                ):
-
-                    def msg(s: str) -> str:
-                        return f'ONNX validation failed for output "{output_name}": {s}'
-
-                    # Due to the presence of top-k operations in the model, the outputs may be
-                    # in different order but still valid. To account for this, we sum
-                    # over the query dimension before comparing.
-                    output_model = output_model.sum(dim=1)
-                    if output_onnx.is_floating_point:
-                        # Convert to fp32 to avoid overflow issues when summing in fp16.
-                        output_onnx = output_onnx.float()
-                    output_onnx = output_onnx.sum(dim=1)
-
-                    if output_model.is_floating_point:
-                        # Absolute and relative tolerances are a bit arbitrary and taken from here:
-                        # https://github.com/pytorch/pytorch/blob/main/torch/onnx/_internal/exporter/_core.py#L1611-L1618
-                        torch.testing.assert_close(
-                            output_onnx,
-                            output_model,
-                            msg=msg,
-                            equal_nan=True,
-                            check_device=False,
-                            check_dtype=False,
-                            check_layout=False,
-                            atol=5e-3,
-                            rtol=1e-1,
-                        )
-                    else:
-                        _torch_testing.assert_most_equal(
-                            output_onnx,
-                            output_model,
-                            msg=msg,
-                        )
-
-        logger.info(f"Successfully exported ONNX model to '{out}'")
-
-    def export_tensorrt(
-        self,
-        out: PathLike,
-        *,
-        precision: Literal["fp32", "fp16"] = "fp32",
-        onnx_args: dict[str, Any] | None = None,
-        max_batchsize: int = 1,
-        opt_batchsize: int = 1,
-        min_batchsize: int = 1,
-        verbose: bool = False,
-    ) -> None:
-        """Build a TensorRT engine from an ONNX model.
-
-        .. note::
-            TensorRT is not part of LightlyTrain’s dependencies and must be installed separately.
-            Installation depends on your OS, Python version, GPU, and NVIDIA driver/CUDA setup.
-            See the [TensorRT documentation](https://docs.nvidia.com/deeplearning/tensorrt/latest/installing-tensorrt/installing.html) for more details.
-            On CUDA 12.x systems you can often install the Python package via `pip install tensorrt-cu12`.
-
-        This loads the ONNX file, parses it with TensorRT, infers the static input
-        shape (C, H, W) from the `"images"` input, and creates an engine with a
-        dynamic batch dimension in the range `[min_batchsize, opt_batchsize, max_batchsize]`.
-        Spatial dimensions must be static in the ONNX model (dynamic H/W are not yet supported).
-
-        The engine is serialized and written to `out`.
-
-        Args:
-            out:
-                Path where the TensorRT engine will be saved.
-            precision:
-                Precision for ONNX export and TensorRT engine building. Either
-                "fp32" or "fp16".
-            onnx_args:
-                Optional arguments to pass to `export_onnx` when exporting
-                the ONNX model prior to building the TensorRT engine. If None,
-                default arguments are used and the ONNX file is saved alongside
-                the TensorRT engine with the same name but `.onnx` extension.
-            max_batchsize:
-                Maximum supported batch size.
-            opt_batchsize:
-                Batch size TensorRT optimizes for.
-            min_batchsize:
-                Minimum supported batch size.
-            verbose:
-                Enable verbose TensorRT logging.
-
-        Raises:
-            FileNotFoundError: If the ONNX file does not exist.
-            RuntimeError: If the ONNX cannot be parsed or engine building fails.
-            ValueError: If batch size constraints are invalid or H/W are dynamic.
-        """
-        model_dtype = next(self.parameters()).dtype
-
-        onnx_args = dict(onnx_args) if onnx_args is not None else {}
-        onnx_args.setdefault("precision", precision)
-
-        tensorrt_helpers.export_tensorrt(
-            export_onnx_fn=self.export_onnx,
-            out=out,
-            precision=precision,
-            model_dtype=model_dtype,
-            onnx_args=onnx_args,
-            max_batchsize=max_batchsize,
-            opt_batchsize=opt_batchsize,
-            min_batchsize=min_batchsize,
-            # We convert the fp32 attention scores already during ONNX export
-            fp32_attention_scores=False,
-            verbose=verbose,
         )
